reducing POS errors to increase HMM accuracy
started converting rdfTypes to database entries to save disk space and improve scalability
moved internet procedures to new class
made writing wordClassCache more definitive - no longer depending on the database lastSourceId to read in words (because we are no longer updating that table when we process sources.  Writing to these tables including the wordrelations table and wordfrequency table are relegated to corpus wide actions that can be done independently of the source processing.  This allows these tables to be curated and source processing to be corrected without corrupting these tables.)
added submenu and decreased scrolling times to allow searching through authors to be much faster.
greatly refined treatment of dashed words.
locked down on words to greatly reduce incorrect words and maximize possibility that words will become defined.
corrected readdate code.  Refined to prevent false positives.
changed mainEntryMap to be an unordered map for performance purposes and also so that a custom comparator is not needed.
tightened rdf lookup requirements to reduce failure on terms