conforming with ST and Longman
Researching each word so, as, etc with Longman has clarified the grammatical roles for these words.  Corrected a bug which failed to remove winner flag from some syntactical descendants.  Removed interjections from those words that can be skipped and still match a pattern because there are too many (it is an open class).