60217/5965465 1.009% corrected setting of minimum cost
corrected setting of minimum cost due to the same PEMA position being repeated in the costing array.  merging all tag sets for ROLE and then sorting them so that the no object shall be evaluated before any object which is contained within it.  This is the fix for _NOUN[9] and other patterns which incorrectly depended on blocking or not to come up with the correct costing, but this was an incorrect approach, as it really depended on the order in which the patterns were evaluated.  Also changed maximum tag time limit to 70 from 100 ms.