concentrates costing in the tFI class
This lowers and targets changes in the cost of forms for words.  If costs are changed during word processing, it makes testing less reproducible because when multiple sources are distributed amongst threads, different sources are procesed in a different order, so that some sources affect the costs of words in sources processed after them, in a way that is not predictable due to timing.